requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
gradio
accelerate
bitsandbytes
blobfile
sentencepiece
accelerate
